fix: add connection header band and use register hook
